refactor: Enhance Supply and Purchase forms and add search to Supplies
This commit introduces significant improvements to the user experience for managing supplies and purchases. It refactors the Supply and Purchase forms for better unit handling and adds search functionality to the supply list.

### Key Features & Enhancements:

-   **Supply Search Functionality:**
    -   A search bar is now implemented on the `SupplyListFragment`, allowing users to filter supplies by name in real time.
    -   `SupplyViewModel` is updated with filtering logic, and the fragment implements `BaseFragment.SearchableFragment`.
    -   The UI now shows an appropriate empty state message when no supplies match the search query.

-   **Improved Supply Form (`SupplyFormFragment`):**
    -   Replaced the free-text input for "Purchase Unit" with a `Spinner`, providing a predefined list of common units (kg, caja, saco, etc.) for consistency.
    -   The UI and logic are streamlined to handle an optional purchase unit and its corresponding conversion factor.

-   **Enhanced Purchase Form (`PurchaseFormFragment`):**
    -   When creating a purchase, the form now dynamically displays a `RadioGroup` allowing the user to select whether the entered quantity is in the supply's base unit (e.g., "gr") or its defined purchase unit (e.g., "kg").
    -   The application automatically handles the conversion to the base unit before saving the purchase, ensuring stock is updated correctly.
    -   The logic for selecting a supply from the spinner has been made more robust.

### Refactoring & Minor Changes:

-   **Code Cleanup:**
    -   Removed commented-out code and unnecessary `println` statements from various files, including `ProductFormFragment`, `RegisterSaleUseCase`, and `HomeViewModel`.
    -   The `RegisterSupplyPurchaseUseCase` was simplified by moving unit conversion logic to the `PurchaseFormFragment`, where the user now explicitly selects the unit.
-   **UI Fixes:**
    -   The "add variant" and "add presentation" buttons on the product form now use a neutral gray background color.
    -   Currency values displayed in the `ProductVariantAdapter` and `ProductPresentationAdapter` now use the standardized `toSoles()` extension for consistent formatting.
    -   Corrected a minor bug in the `ProductAdapter` where the price was being displayed with a `$` prefix instead of the `toSoles()` formatted string.
-   **Database:**
    -   The `SaleDao` queries for fetching sales by date range have been corrected to use `strftime` to compare only the date part, ensuring accurate filtering regardless of the time.

+++ app/src/main/java/com/rige/dulcegest/domain/usecases/home/CalculateWeeklyNetProfitUseCase.kt
@@ -54,7 +54,7 @@ class CalculateWeeklyNetProfitUseCase @Inject constructor(
 
         val totalCostSum: Double = lastBatches.sumOf { it.totalCost }
 
-        val totalQuantitySum: Double = lastBatches.sumOf { it.quantityProduced.toDouble() }
+        val totalQuantitySum: Double = lastBatches.sumOf { it.quantityProduced }
 
         if (totalQuantitySum == 0.0) {
             return 0.0
